Implementation of SetUniqueID API
Implemented SetUniqueID API which will allow client to change the
uniqueID, if possible. In case it is not possible to set the provided id
as the new id, false will be returned.
Also while parsing a buffer an automatic try is made to use the id
present in the buffer be used as the uniqueID, if possible.